Lwm2m: backEnd: refactoring security
