+++ common/transport/lwm2m/src/main/java/org/thingsboard/server/transport/lwm2m/server/LwM2MTransportRequest.java
@@ -86,7 +86,7 @@ public class LwM2MTransportRequest {
     private LwM2mValueConverterImpl converter;
 
     @Autowired
-    LwM2MTransportService service;
+    LwM2MTransportServiceImpl service;
 
 
     @PostConstruct
