+++ common/transport/lwm2m/src/main/java/org/thingsboard/server/transport/lwm2m/server/secure/LwM2mInMemorySecurityStore.java
@@ -25,32 +25,32 @@ import org.eclipse.leshan.server.security.SecurityInfo;
 import org.eclipse.leshan.server.security.SecurityStoreListener;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import org.thingsboard.server.common.data.DeviceProfile;
 import org.thingsboard.server.gen.transport.TransportProtos;
-import org.thingsboard.server.transport.lwm2m.secure.LwM2MGetSecurityInfo;
+import org.thingsboard.server.transport.lwm2m.secure.LwM2MSecurityMode;
+import org.thingsboard.server.transport.lwm2m.secure.LwM2mValidateCredentialsSecurityInfo;
 import org.thingsboard.server.transport.lwm2m.secure.ReadResultSecurityStore;
 import org.thingsboard.server.transport.lwm2m.server.LwM2MTransportHandler;
 import org.thingsboard.server.transport.lwm2m.server.client.AttrTelemetryObserveValue;
 import org.thingsboard.server.transport.lwm2m.server.client.LwM2MClient;
 import org.thingsboard.server.transport.lwm2m.utils.TypeServer;
 
-import java.util.Map;
-import java.util.UUID;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
+import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReadWriteLock;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 import java.util.stream.Collectors;
 
-import static org.thingsboard.server.transport.lwm2m.secure.LwM2MSecurityMode.DEFAULT_MODE;
 import static org.thingsboard.server.transport.lwm2m.secure.LwM2MSecurityMode.NO_SEC;
 
 @Slf4j
-@Component("LwM2mInMemorySecurityStore")
+@Service("LwM2mInMemorySecurityStore")
 @ConditionalOnExpression("('${service.type:null}'=='tb-transport' && '${transport.lwm2m.enabled:false}'=='true' )|| ('${service.type:null}'=='monolith' && '${transport.lwm2m.enabled}'=='true')")
 public class LwM2mInMemorySecurityStore extends InMemorySecurityStore {
     // lock for the two maps
@@ -63,26 +63,36 @@ public class LwM2mInMemorySecurityStore extends InMemorySecurityStore {
     private SecurityStoreListener listener;
 
     @Autowired
-    LwM2MGetSecurityInfo lwM2MGetSecurityInfo;
+    LwM2mValidateCredentialsSecurityInfo lwM2MValidateCredentialsSecurityInfo;
 
+    /**
+     * Start after DefaultAuthorizer or LwM2mPskStore
+     * @param endPoint -
+     * @return SecurityInfo
+     */
     @Override
     public SecurityInfo getByEndpoint(String endPoint) {
         readLock.lock();
         try {
             String registrationId = this.getByRegistrationId(endPoint, null);
-            SecurityInfo info = (registrationId != null && sessions.size() > 0 && sessions.get(registrationId) != null) ? sessions.get(registrationId).getInfo() : this.add(endPoint);
+            SecurityInfo info = (registrationId != null && sessions.size() > 0 && sessions.get(registrationId) != null) ? sessions.get(registrationId).getInfo() : this.addLwM2MClientToSession(endPoint);
             return info;
         } finally {
             readLock.unlock();
         }
     }
 
+    /**
+     * Start after LwM2mPskStore
+     * @param identity -
+     * @return SecurityInfo
+     */
     @Override
     public SecurityInfo getByIdentity(String identity) {
         readLock.lock();
         try {
             String integrationId = this.getByRegistrationId(null, identity);
-            return (integrationId != null) ? sessions.get(integrationId).getInfo() : add(identity);
+            return (integrationId != null) ? sessions.get(integrationId).getInfo() : this.addLwM2MClientToSession(identity);
         } finally {
             readLock.unlock();
         }
@@ -141,11 +151,19 @@ public class LwM2mInMemorySecurityStore extends InMemorySecurityStore {
         return this.getSession(new UUID(sessionInfo.getSessionIdMSB(), sessionInfo.getSessionIdLSB())).entrySet().iterator().next().getValue();
 
     }
+
+    /**
+     * Update in sessions (LwM2MClient for key registration_Id) after starting registration LwM2MClient in LwM2MTransportServiceImpl
+     * Remove from sessions LwM2MClient with key registration_Endpoint
+     * @param lwServer -
+     * @param registration -
+     * @return LwM2MClient after adding it to session
+     */
     public LwM2MClient updateInSessionsLwM2MClient(LeshanServer lwServer, Registration registration) {
         writeLock.lock();
         try {
             if (this.sessions.get(registration.getEndpoint()) == null) {
-                this.add(registration.getEndpoint());
+                this.addLwM2MClientToSession(registration.getEndpoint());
             }
             LwM2MClient lwM2MClient = this.sessions.get(registration.getEndpoint());
             lwM2MClient.setLwServer(lwServer);
@@ -166,45 +184,43 @@ public class LwM2mInMemorySecurityStore extends InMemorySecurityStore {
         return (registrationIds != null && registrationIds.size() > 0) ? registrationIds.get(0) : null;
     }
 
-    public String getByRegistrationId(String credentialsId) {
-        List<String> registrationIds = (this.sessions.entrySet().stream().filter(model -> credentialsId.equals(model.getValue().getEndPoint())).map(model -> model.getKey()).collect(Collectors.toList()).size() > 0) ?
-                this.sessions.entrySet().stream().filter(model -> credentialsId.equals(model.getValue().getEndPoint())).map(model -> model.getKey()).collect(Collectors.toList()) :
-                this.sessions.entrySet().stream().filter(model -> credentialsId.equals(model.getValue().getIdentity())).map(model -> model.getKey()).collect(Collectors.toList());
-        return (registrationIds != null && registrationIds.size() > 0) ? registrationIds.get(0) : null;
-    }
-
     public Registration getByRegistration(String registrationId) {
         return this.sessions.get(registrationId).getRegistration();
     }
 
-    private SecurityInfo add(String identity) {
-        ReadResultSecurityStore store = lwM2MGetSecurityInfo.getSecurityInfo(identity, TypeServer.CLIENT);
-        UUID profileUuid = (store.getDeviceProfile() != null && addUpdateProfileParameters(store.getDeviceProfile())) ? store.getDeviceProfile().getUuidId() : null;
-        if (store.getSecurityInfo() != null) {
-            if (store.getSecurityMode() < DEFAULT_MODE.code) {
+    /**
+     * Add new LwM2MClient to session
+     * @param identity-
+     * @return SecurityInfo. If error - SecurityInfoError
+     * and log:
+     * - FORBIDDEN - if there is no authorization
+     * - profileUuid - if the device does not have a profile
+     * - device - if the thingsboard does not have a device with a name equal to the identity
+     */
+    private SecurityInfo addLwM2MClientToSession(String identity) {
+        ReadResultSecurityStore store = lwM2MValidateCredentialsSecurityInfo.validateCredentialsSecurityInfo(identity, TypeServer.CLIENT);
+        if (store.getSecurityMode() < LwM2MSecurityMode.DEFAULT_MODE.code) {
+            UUID profileUuid = (store.getDeviceProfile() != null && addUpdateProfileParameters(store.getDeviceProfile())) ? store.getDeviceProfile().getUuidId() : null;
+            if (store.getSecurityInfo() != null && profileUuid != null) {
                 String endpoint = store.getSecurityInfo().getEndpoint();
-//                sessions.put(endpoint, new LwM2MClient(endpoint, store.getSecurityInfo().getIdentity(), store.getSecurityInfo(), store.getMsg(), null, null, profileUuid));
                 sessions.put(endpoint, new LwM2MClient(endpoint, store.getSecurityInfo().getIdentity(), store.getSecurityInfo(), store.getMsg(), null, profileUuid));
-            }
-        } else {
-            if (store.getSecurityMode() == NO_SEC.code && profileUuid != null)
-//                sessions.put(identity, new LwM2MClient(identity, null, null, store.getMsg(), null, null, profileUuid));
+            } else if (store.getSecurityMode() == NO_SEC.code && profileUuid != null) {
                 sessions.put(identity, new LwM2MClient(identity, null, null, store.getMsg(), null, profileUuid));
-            else {
-                log.error("Registration failed: FORBIDDEN/profileUuid/device [{}] , endpointId: [{}]", profileUuid, identity);
-                /**
-                 * Return Error securityInfo
-                 */
-                byte[] preSharedKey = Hex.decodeHex("0A0B".toCharArray());
-                SecurityInfo info = SecurityInfo.newPreSharedKeyInfo("error", "error_identity", preSharedKey);
-                return info;
+            } else {
+                    log.error("Registration failed: FORBIDDEN/profileUuid/device [{}] , endpointId: [{}]", profileUuid, identity);
+                    /**
+                     * Return Error securityInfo
+                     */
+                    byte[] preSharedKey = Hex.decodeHex("0A0B".toCharArray());
+                    SecurityInfo infoError = SecurityInfo.newPreSharedKeyInfo("error", "error_identity", preSharedKey);
+                    return infoError;
+                }
             }
-        }
-        return store.getSecurityInfo();
+            return store.getSecurityInfo();
     }
 
-    public Map<String, LwM2MClient> getSession (UUID sessionUuId){
-       return this.sessions.entrySet().stream().filter(e -> e.getValue().getSessionUuid().equals(sessionUuId)).collect(Collectors.toMap(map -> map.getKey(), map -> map.getValue()));
+    public Map<String, LwM2MClient> getSession(UUID sessionUuId) {
+        return this.sessions.entrySet().stream().filter(e -> e.getValue().getSessionUuid().equals(sessionUuId)).collect(Collectors.toMap(map -> map.getKey(), map -> map.getValue()));
     }
 
     public Map<String, LwM2MClient> getSessions() {
@@ -219,13 +235,10 @@ public class LwM2mInMemorySecurityStore extends InMemorySecurityStore {
         return this.profiles.get(profileUuId);
     }
 
-    public Map<UUID, AttrTelemetryObserveValue>setProfiles(Map<UUID, AttrTelemetryObserveValue> profiles) {
+    public Map<UUID, AttrTelemetryObserveValue> setProfiles(Map<UUID, AttrTelemetryObserveValue> profiles) {
         return this.profiles = profiles;
     }
 
-    /**
-     * @param deviceProfile
-     */
     public boolean addUpdateProfileParameters(DeviceProfile deviceProfile) {
         JsonObject profilesConfigData = LwM2MTransportHandler.getObserveAttrTelemetryFromThingsboard(deviceProfile);
         if (profilesConfigData != null) {
@@ -233,5 +246,4 @@ public class LwM2mInMemorySecurityStore extends InMemorySecurityStore {
         }
         return (profilesConfigData != null);
     }
-
 }
